Fix release build crash.
When compiling a release build the compiler generated
code which used ES segment selector. That was set
to a null selector so running the code made
a crash.

This patch also sets FS and GS segment selectors to
contain valid values to avoid the same problem as with
the ES segment selector.